+++ FrontEnd/Hospitality 2/Hospitality/Views/Patient/Appointment_History/Appointment_Detail_View.swift
@@ -1,5 +1,46 @@
 import SwiftUI
 
+// MARK: - AppointmentDetailViewModel (moved to top level)
+class AppointmentDetailViewModel: ObservableObject {
+    // Appointment details with diagnosis and prescription
+    @Published var appointmentDetails: AppointmentDetailResponse?
+    @Published var isLoadingDetails = false
+    @Published var detailsError: String?
+    
+    private let doctorServices = DoctorServices()
+    
+    func loadAllData(appointment: DoctorResponse.DocAppointment) {
+        loadAppointmentDetails(appointmentId: appointment.appointmentId)
+    }
+    
+    func loadAppointmentDetails(appointmentId: Int) {
+        isLoadingDetails = true
+        detailsError = nil
+        print("🔍 Loading appointment details for appointment ID: \(appointmentId)")
+        
+        Task {
+            do {
+                let details = try await doctorServices.fetchAppointmentDetails(appointmentId: appointmentId)
+                DispatchQueue.main.async {
+                    self.appointmentDetails = details
+                    self.isLoadingDetails = false
+                    print("✅ Successfully loaded appointment details: \(details)")
+                    print("Diagnosis data: \(String(describing: details.diagnosis))")
+                    if let diagnosisData = details.diagnosis?.diagnosisData {
+                        print("Diagnosis data items: \(diagnosisData)")
+                    }
+                }
+            } catch {
+                DispatchQueue.main.async {
+                    self.detailsError = error.localizedDescription
+                    self.isLoadingDetails = false
+                    print("❌ Failed to load appointment details: \(error)")
+                }
+            }
+        }
+    }
+}
+
 // MARK: - Appointment Detail View
 struct AppointmentDetailView: View {
     // MARK: - Properties
@@ -20,47 +61,6 @@ struct AppointmentDetailView: View {
                     await loadAppointmentDetails()
                 }
             }
-    }
-    
-    // MARK: - AppointmentDetailViewModel
-    class AppointmentDetailViewModel: ObservableObject {
-        // Appointment details with diagnosis and prescription
-        @Published var appointmentDetails: AppointmentDetailResponse?
-        @Published var isLoadingDetails = false
-        @Published var detailsError: String?
-        
-        private let doctorServices = DoctorServices()
-        
-        func loadAllData(appointment: DoctorResponse.DocAppointment) {
-            loadAppointmentDetails(appointmentId: appointment.appointmentId)
-        }
-        
-        func loadAppointmentDetails(appointmentId: Int) {
-            isLoadingDetails = true
-            detailsError = nil
-            print("🔍 Loading appointment details for appointment ID: \(appointmentId)")
-            
-            Task {
-                do {
-                    let details = try await doctorServices.fetchAppointmentDetails(appointmentId: appointmentId)
-                    DispatchQueue.main.async {
-                        self.appointmentDetails = details
-                        self.isLoadingDetails = false
-                        print("✅ Successfully loaded appointment details: \(details)")
-                        print("Diagnosis data: \(String(describing: details.diagnosis))")
-                        if let diagnosisData = details.diagnosis?.diagnosisData {
-                            print("Diagnosis data items: \(diagnosisData)")
-                        }
-                    }
-                } catch {
-                    DispatchQueue.main.async {
-                        self.detailsError = error.localizedDescription
-                        self.isLoadingDetails = false
-                        print("❌ Failed to load appointment details: \(error)")
-                    }
-                }
-            }
-        }
     }
     
     // MARK: - StatusBadge
@@ -226,7 +226,7 @@ struct AppointmentDetailView: View {
     @ViewBuilder
     private func overviewContent(_ appointment: AppointmentDetailResponse) -> some View {
         // Appointment overview card
-        AppointmentOverviewCard(appointment: appointment)
+        AppointmentOverviewCard(appointment: appointment, viewModel: viewModel)
 
         // Summary cards if available
         if appointment.prescription != nil || appointment.diagnosis != nil {
@@ -542,6 +542,8 @@ struct SummaryCard_Previews: PreviewProvider {
 // MARK: - Appointment Overview Card
 struct AppointmentOverviewCard: View {
     let appointment: AppointmentDetailResponse
+    @State private var showRescheduleView = false
+    @ObservedObject var viewModel: AppointmentDetailViewModel
     
     var body: some View {
         VStack(alignment: .leading, spacing: 16) {
@@ -554,39 +556,53 @@ struct AppointmentOverviewCard: View {
                 if let reason = appointment.reason, !reason.isEmpty {
                     InfoRow(label: "Reason", value: reason)
                 }
+                
+                // Add reschedule button only for upcoming appointments
+                if appointment.status.lowercased() == "upcoming" || 
+                   appointment.status.lowercased() == "scheduled" {
+                    Button(action: {
+                        showRescheduleView = true
+                    }) {
+                        HStack {
+                            Image(systemName: "calendar.badge.clock")
+                            Text("Reschedule Appointment")
+                            Spacer()
+                            Image(systemName: "chevron.right")
+                                .font(.caption)
+                        }
+                        .padding()
+                        .background(Color.blue.opacity(0.1))
+                        .foregroundColor(.blue)
+                        .cornerRadius(8)
+                    }
+                    .padding(.top, 8)
+                }
             }
             .padding()
             .background(Color(.secondarySystemBackground))
             .cornerRadius(12)
         }
+        .sheet(isPresented: $showRescheduleView) {
+            AppointmentRescheduleView(
+                appointmentId: appointment.appointmentId,
+                doctorId: String(appointment.staffId),
+                currentDate: appointment.date,
+                currentSlotId: appointment.slotId,
+                reason: appointment.reason ?? "",
+                onRescheduleComplete: {
+                    showRescheduleView = false
+                    // Refresh the appointment details
+                    viewModel.loadAppointmentDetails(appointmentId: appointment.appointmentId)
+                }
+            )
+        }
     }
     
     private func formatDate(_ dateString: String) -> String {
-        // First try the ISO format with time
-        let isoFormatter = DateFormatter()
-        isoFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        
-        if let date = isoFormatter.date(from: dateString) {
-            let displayFormatter = DateFormatter()
-            displayFormatter.dateStyle = .medium
-            displayFormatter.timeStyle = .short
-            return displayFormatter.string(from: date)
-        }
-        
-        // Then try simple date format
-        let simpleFormatter = DateFormatter()
-        simpleFormatter.dateFormat = "yyyy-MM-dd"
-        
-        if let date = simpleFormatter.date(from: dateString) {
-            let displayFormatter = DateFormatter()
-            displayFormatter.dateStyle = .medium
-            displayFormatter.timeStyle = .none
-            return displayFormatter.string(from: date)
-        }
-        
-        // Return original if we can't parse it
+        // Existing date formatting code...
         return dateString
-    }}
+    }
+}
 
 // MARK: - Info Row
 struct InfoRow: View {
