+++ FrontEnd/Hospitality 2/Hospitality/Views/Patient/DoctorRecommeder/SymptomViewModel.swift
@@ -1,199 +1,178 @@
-import Foundation
-import Combine
-import SwiftUI
-
-class SymptomViewModel: ObservableObject {
-    struct Question {
-        let text: String
-        let options: [String]
-        let allowMultiple: Bool
-        
-        init(text: String, options: [String], allowMultiple: Bool = false) {
-            self.text = text
-            self.options = options
-            self.allowMultiple = allowMultiple
-        }
-    }
+// struct AppointmentHistoryCard: View {
+//     let appointment: PatientAppointHistoryListResponse
+//     @Environment(\.colorScheme) var colorScheme
+//     @State private var slotStartTime: String = "Loading..."
+//     @State private var isSlotLoading: Bool = false
+//     @State private var staffName: String = "Loading..."
+//     @State private var isStaffNameLoading: Bool = false
     
-    // Questions based on CSV structure - now with multi-selection support
-    let questions = [
-        Question(
-            text: "Which part of your body is affected or feels uncomfortable?",
-            options: ["Head", "Chest", "Stomach/Abdomen", "Back/Spine", "Limbs (Arms/Legs)", "Skin/Hair", "Ear/Nose/Throat", "Genitals/Urinary System", "Whole Body"],
-            allowMultiple: true
-        ),
-        Question(
-            text: "What best describes your main symptom?",
-            options: ["Pain", "Swelling", "Rash or skin issue", "Bleeding", "Movement difficulty", "Mental health issue", "Digestive problems", "Breathing problems", "Heart-related symptoms"],
-            allowMultiple: true
-        ),
-        Question(
-            text: "How long have you had this issue?",
-            options: ["Just today", "1–3 days", "4–7 days", "More than a week", "Keeps coming back"]
-        ),
-        Question(
-            text: "How would you describe the severity?",
-            options: ["Mild", "Moderate", "Severe (disrupting daily life)", "Emergency"]
-        ),
-        Question(
-            text: "What is your age group?",
-            options: ["Child (0-12 years)", "Teenager (13-19 years)", "Adult (20-60 years)", "Senior (60+ years)"]
-        ),
-        Question(
-            text: "Do you have any of these conditions?",
-            options: ["Diabetes or blood sugar issues", "Thyroid problems", "High blood pressure", "Heart condition", "Cancer history", "None of these"],
-            allowMultiple: true
-        ),
-        Question(
-            text: "Are you experiencing any of these additional symptoms?",
-            options: ["Fever", "Nausea/Vomiting", "Dizziness", "Fatigue", "Weight change", "Vision changes", "None of these"],
-            allowMultiple: true
-        ),
-//        Question(
-//            text: "For female patients: Is this related to reproductive health?",
-//            options: ["Yes - pregnancy related", "Yes - menstrual related", "Yes - other reproductive issue", "Not applicable", "No"]
-//        )
-    ]
-    
-    @Published var currentQuestionIndex = 0
-    @Published var multiAnswers: [[String]] = []
-    @Published var additionalNotes: String = ""
-    @Published var recommendation: DoctorRecommendation?
-    @Published var isLoading = false
-    @Published var isShowingResult = false
-    @Published var showErrorMessage = false
-    @Published var errorMessage = ""
-    
-    private var cancellables = Set<AnyCancellable>()
-    private let geminiService = GeminiService()
-    
-    var currentQuestion: Question {
-        questions[currentQuestionIndex]
-    }
-    
-    init() {
-        // Initialize the multiAnswers array with empty arrays
-        multiAnswers = Array(repeating: [], count: questions.count)
-    }
-    
-    func toggleAnswer(_ answer: String) {
-        if currentQuestion.allowMultiple {
-            // For multi-select questions, toggle the presence of the answer
-            if multiAnswers[currentQuestionIndex].contains(answer) {
-                // Remove answer if already selected
-                multiAnswers[currentQuestionIndex].removeAll { $0 == answer }
-                
-                // If "None of these" is selected, deselect everything else
-                if answer == "None of these" {
-                    multiAnswers[currentQuestionIndex] = []
-                }
-            } else {
-                // Add answer if not already selected
-                
-                // If selecting "None of these", clear other selections
-                if answer == "None of these" {
-                    multiAnswers[currentQuestionIndex] = ["None of these"]
-                } else {
-                    // If selecting something else, remove "None of these" if present
-                    if multiAnswers[currentQuestionIndex].contains("None of these") {
-                        multiAnswers[currentQuestionIndex].removeAll { $0 == "None of these" }
-                    }
+//     var body: some View {
+//         VStack(alignment: .leading, spacing: 12) {
+//             HStack {
+//                 VStack(alignment: .leading, spacing: 4) {
+//                     Text(staffName)
+//                         .font(.system(size: 18, weight: .semibold, design: .rounded))
                     
-                    multiAnswers[currentQuestionIndex].append(answer)
-                }
-            }
-        } else {
-            // For single-select questions, replace the answer
-            multiAnswers[currentQuestionIndex] = [answer]
+//                     Text("\(formatDate(appointment.date)) • \(slotStartTime)")
+//                         .font(.system(size: 14, weight: .medium, design: .rounded))
+//                         .foregroundColor(colorScheme == .dark ? .white.opacity(0.7) : .gray)
+//                 }
+                
+//                 Spacer()
+                
+//                 StatusBadge(status: appointment.status)
+//             }
             
-            // Move to the next question automatically for single-select questions
-            nextQuestion()
-        }
-    }
+//             Divider()
+            
+//             HStack {
+//                 VStack(alignment: .leading, spacing: 4) {
+//                     Text("Doctor")
+//                         .font(.caption)
+//                         .foregroundColor(colorScheme == .dark ? .white.opacity(0.5) : .gray.opacity(0.7))
+                    
+//                     Text(staffName)
+//                         .font(.system(size: 14, weight: .medium, design: .rounded))
+//                         .foregroundColor(colorScheme == .dark ? .white.opacity(0.9) : .black)
+//                 }
+                
+//                 Spacer()
+                
+//                 VStack(alignment: .trailing, spacing: 4) {
+//                     Text("Time")
+//                         .font(.caption)
+//                         .foregroundColor(colorScheme == .dark ? .white.opacity(0.5) : .gray.opacity(0.7))
+                    
+//                     Text(slotStartTime)
+//                         .font(.system(size: 14, weight: .medium, design: .rounded))
+//                         .foregroundColor(colorScheme == .dark ? .white.opacity(0.9) : .black)
+//                 }
+//             }
+            
+//             if let reason = appointment.reason, !reason.isEmpty {
+//                 Text("Reason: \(reason)")
+//                     .font(.system(size: 14, weight: .regular, design: .rounded))
+//                     .foregroundColor(colorScheme == .dark ? .white.opacity(0.7) : .gray)
+//                     .padding(.top, 4)
+//             } else {
+//                 Text("No reason provided")
+//                     .font(.system(size: 14, weight: .regular, design: .rounded))
+//                     .foregroundColor(colorScheme == .dark ? .white.opacity(0.5) : .gray.opacity(0.7))
+//                     .italic()
+//                     .padding(.top, 4)
+//             }
+//         }
+//         .padding()
+//         .background(
+//             RoundedRectangle(cornerRadius: 12)
+//                 .fill(colorScheme == .dark ? Color(hex: "1E2533") : .white)
+//                 .shadow(
+//                     color: colorScheme == .dark ? Color.black.opacity(0.4) : Color.gray.opacity(0.2),
+//                     radius: 5, x: 0, y: 2
+//                 )
+//         )
+//         .overlay(
+//             RoundedRectangle(cornerRadius: 12)
+//                 .stroke(Color.blue, lineWidth: 1)
+//         )
+//         .onAppear {
+//             loadSlotTime()
+//             loadStaffName()
+//         }
+//     }
     
-    func isAnswerSelected(_ answer: String) -> Bool {
-        return multiAnswers[currentQuestionIndex].contains(answer)
-    }
-    
-    func nextQuestion() {
-        // Only proceed if at least one option is selected for the current question
-        if !multiAnswers[currentQuestionIndex].isEmpty {
-            if currentQuestionIndex < questions.count - 1 {
-                currentQuestionIndex += 1
-            } else {
-                getRecommendation()
-            }
-        }
-    }
-    
-    func goToPreviousQuestion() {
-        if currentQuestionIndex > 0 {
-            currentQuestionIndex -= 1
-        }
-    }
-    
-    func restart() {
-        currentQuestionIndex = 0
-        multiAnswers = Array(repeating: [], count: questions.count)
-        additionalNotes = ""
-        recommendation = nil
-        isShowingResult = false
-    }
-    
-    var urgencyColor: Color {
-        guard let recommendation = recommendation else { return .gray }
-        switch recommendation.urgencyLevel {
-        case 1: return .green
-        case 2: return .orange
-        case 3: return .red
-        default: return .gray
-        }
-    }
-    
-    var urgencyText: String {
-        guard let recommendation = recommendation else { return "Unknown" }
-        switch recommendation.urgencyLevel {
-        case 1: return "Routine"
-        case 2: return "Soon"
-        case 3: return "Urgent"
-        default: return "Unknown"
-        }
-    }
-    
-    func getRecommendation() {
-        isLoading = true
+//     private func formatDate(_ dateString: String) -> String {
+//         let dateFormatter = DateFormatter()
+//         dateFormatter.dateFormat = "yyyy-MM-dd"
         
-        // Prepare formatted symptoms for API call
-        var formattedSymptoms: [String] = []
-        for i in 0..<questions.count {
-            if multiAnswers[i].isEmpty {
-                formattedSymptoms.append("\(questions[i].text): No selection")
-            } else if multiAnswers[i].count == 1 {
-                formattedSymptoms.append("\(questions[i].text): \(multiAnswers[i][0])")
-            } else {
-                let answersJoined = multiAnswers[i].joined(separator: ", ")
-                formattedSymptoms.append("\(questions[i].text): \(answersJoined)")
-            }
-        }
+//         guard let date = dateFormatter.date(from: dateString) else {
+//             return dateString
+//         }
         
-        // Include additional notes if provided
-        if !additionalNotes.isEmpty {
-            formattedSymptoms.append("Additional information: \(additionalNotes)")
-        }
+//         dateFormatter.dateStyle = .medium
+//         return dateFormatter.string(from: date)
+//     }
+    
+//     private func loadSlotTime() {
+//         guard !isSlotLoading else { return }
         
-        geminiService.getRecommendation(symptoms: formattedSymptoms)
-            .receive(on: DispatchQueue.main)
-            .sink(receiveCompletion: { [weak self] completion in
-                if case .failure(let error) = completion {
-                    self?.isLoading = false
-                    self?.errorMessage = error.localizedDescription
-                    self?.showErrorMessage = true
-                }
-            }, receiveValue: { [weak self] recommendation in
-                self?.recommendation = recommendation
-                self?.isLoading = false
-                self?.isShowingResult = true
-            })
-            .store(in: &cancellables)
-    }
-}
+//         isSlotLoading = true
+        
+//         Task {
+//             do {
+//                 let dateFormatter = DateFormatter()
+//                 dateFormatter.dateFormat = "yyyy-MM-dd"
+//                 guard let appointmentDate = dateFormatter.date(from: String(appointment.date.prefix(10))) else {
+//                     throw NetworkError.unknownError
+//                 }
+                
+//                 let dateString = dateFormatter.string(from: appointmentDate)
+//                 let slots = try await DoctorServices().fetchDoctorSlots(doctorId: appointment.staff_id ?? "", date: dateString)
+                
+//                 if let slot = slots.first(where: { $0.slot_id == appointment.slot_id }) {
+//                     let timeFormatter = DateFormatter()
+//                     timeFormatter.dateFormat = "HH:mm:ss"
+//                     if let timeDate = timeFormatter.date(from: slot.slot_start_time) {
+//                         timeFormatter.dateFormat = "HH:mm"
+//                         DispatchQueue.main.async {
+//                             slotStartTime = timeFormatter.string(from: timeDate)
+//                             isSlotLoading = false
+//                         }
+//                     } else {
+//                         DispatchQueue.main.async {
+//                             slotStartTime = "N/A"
+//                             isSlotLoading = false
+//                         }
+//                     }
+//                 } else {
+//                     DispatchQueue.main.async {
+//                         slotStartTime = "N/A"
+//                         isSlotLoading = false
+//                     }
+//                 }
+//             } catch {
+//                 DispatchQueue.main.async {
+//                     slotStartTime = "N/A"
+//                     isSlotLoading = false
+//                 }
+//             }
+//         }
+//     }
+    
+//     private func loadStaffName() {
+//         guard !isStaffNameLoading else { return }
+        
+//         isStaffNameLoading = true
+        
+//         Task {
+//             do {
+//                 guard let url = URL(string: "\(Constants.baseURL)/hospital/general/doctors/\(appointment.staff_id ?? "")/") else {
+//                     throw NetworkError.invalidURL
+//                 }
+                
+//                 var request = URLRequest(url: url)
+//                 request.httpMethod = "GET"
+//                 request.addValue("application/json", forHTTPHeaderField: "Content-Type")
+//                 request.addValue("Bearer \(UserDefaults.accessToken)", forHTTPHeaderField: "Authorization")
+                
+//                 let (data, response) = try await URLSession.shared.data(for: request)
+                
+//                 guard let httpResponse = response as? HTTPURLResponse, (200...299).contains(httpResponse.statusCode) else {
+//                     throw NetworkError.invalidResponse
+//                 }
+                
+//                 let doctor = try JSONDecoder().decode(PatientSpecificDoctorResponse.self, from: data)
+                
+//                 DispatchQueue.main.async {
+//                     staffName = doctor.staff_name
+//                     isStaffNameLoading = false
+//                 }
+//             } catch {
+//                 DispatchQueue.main.async {
+//                     staffName = "Unknown Doctor"
+//                     isStaffNameLoading = false
+//                 }
+//             }
+//         }
+//     }
+// }
