+++ FrontEnd/Hospitality 2/Hospitality/Views/Patient/HomePatient.swift
@@ -833,390 +833,6 @@ struct AppointmentDetailOverlay: View {
 
 
 
-// MARK: - AppointmentRescheduleView
-struct AppointmentRescheduleView: View {
-    let appointmentId: Int
-    let doctorId: String
-    let currentDate: String
-    let currentSlotId: Int
-    let reason: String
-    let onRescheduleComplete: () -> Void
-    
-    @Environment(\.presentationMode) var presentationMode
-    @Environment(\.colorScheme) var colorScheme
-    @State private var doctor: PatientSpecificDoctorResponse?
-    @State private var slots: [PatientSlotListResponse] = []
-    @State private var selectedDate = Date()
-    @State private var selectedSlot: PatientSlotListResponse?
-    @State private var isLoading = false
-    @State private var errorMessage: String?
-    @State private var rescheduleSuccess = false
-    @State private var showConfirmation = false
-    
-    private var initialDate: Date {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd"
-        return formatter.date(from: currentDate) ?? Date()
-    }
-    
-    var body: some View {
-        NavigationView {
-            ScrollView {
-                VStack(spacing: 20) {
-                    if isLoading && doctor == nil {
-                        ProgressView()
-                            .frame(maxWidth: .infinity, minHeight: 300)
-                    } else if let errorMessage = errorMessage {
-                        ErrorView(message: errorMessage, onRetry: fetchDoctorDetails)
-                    } else {
-                        if let doctor = doctor {
-                            VStack(spacing: 16) {
-                                HStack(spacing: 16) {
-                                    ZStack {
-                                        Circle()
-                                            .fill(Color.blue.opacity(0.1))
-                                            .frame(width: 80, height: 80)
-                                        
-                                        Image(systemName: "person.fill")
-                                            .font(.system(size: 40))
-                                            .foregroundColor(.blue)
-                                    }
-                                    
-                                    VStack(alignment: .leading, spacing: 6) {
-                                        Text(doctor.staff_name)
-                                            .font(.title2)
-                                            .fontWeight(.bold)
-                                        
-                                        Text(doctor.specialization)
-                                            .font(.subheadline)
-                                            .foregroundColor(.secondary)
-                                    }
-                                    
-                                    Spacer()
-                                }
-                            }
-                            .padding()
-                            .background(Color(.systemBackground))
-                            .cornerRadius(12)
-                            .shadow(color: Color.black.opacity(0.1), radius: 5, x: 0, y: 2)
-                        }
-                        
-                        VStack(alignment: .leading, spacing: 12) {
-                            Text("Current Appointment")
-                                .font(.headline)
-                            
-                            HStack {
-                                VStack(alignment: .leading, spacing: 8) {
-                                    Text("Date: \(formatDate(currentDate))")
-                                    Text("Appointment ID: \(appointmentId)")
-                                    if !reason.isEmpty {
-                                        Text("Reason: \(reason)")
-                                    }
-                                }
-                                Spacer()
-                            }
-                            .padding()
-                            .background(Color(.secondarySystemBackground))
-                            .cornerRadius(8)
-                        }
-                        .padding()
-                        .background(Color(.systemBackground))
-                        .cornerRadius(12)
-                        .shadow(color: Color.black.opacity(0.1), radius: 5, x: 0, y: 2)
-                        
-                        VStack(alignment: .leading, spacing: 12) {
-                            Text("Select New Date")
-                                .font(.headline)
-                            
-                            DatePicker("",
-                                       selection: $selectedDate,
-                                       in: Date()...(Calendar.current.date(byAdding: .month, value: 3, to: Date()) ?? Date()),
-                                       displayedComponents: .date)
-                            .datePickerStyle(.graphical)
-                            .onChange(of: selectedDate) { _ in
-                                fetchSlots()
-                            }
-                        }
-                        .padding()
-                        .background(Color(.systemBackground))
-                        .cornerRadius(12)
-                        .shadow(color: Color.black.opacity(0.1), radius: 5, x: 0, y: 2)
-                        
-                        if isLoading && slots.isEmpty {
-                            ProgressView()
-                                .frame(maxWidth: .infinity, minHeight: 100)
-                        } else if slots.isEmpty {
-                            Text("No available slots for selected date")
-                                .foregroundColor(.secondary)
-                                .padding()
-                        } else {
-                            VStack(alignment: .leading, spacing: 12) {
-                                Text("Available Slots")
-                                    .font(.headline)
-                                
-                                LazyVGrid(columns: [GridItem(.adaptive(minimum: 100), spacing: 10)], spacing: 10) {
-                                    ForEach(slots, id: \.slot_id) { slot in
-                                        SlotButton(
-                                            slot: slot,
-                                            isSelected: selectedSlot?.slot_id == slot.slot_id,
-                                            onSelect: {
-                                                if !isSlotPassed(date: selectedDate, timeString: slot.slot_start_time) {
-                                                    selectedSlot = slot
-                                                }
-                                            },
-                                            isPassed: isSlotPassed(date: selectedDate, timeString: slot.slot_start_time)
-                                        )
-                                    }
-                                }
-                            }
-                            .padding()
-                            .background(Color(.systemBackground))
-                            .cornerRadius(12)
-                            .shadow(color: Color.black.opacity(0.1), radius: 5, x: 0, y: 2)
-                        }
-                        
-                        Button(action: rescheduleAppointment) {
-                            Text("Reschedule Appointment")
-                                .font(.headline)
-                                .foregroundColor(.white)
-                                .frame(maxWidth: .infinity)
-                                .padding()
-                                .background(selectedSlot == nil ? Color.gray : Color.blue)
-                                .cornerRadius(10)
-                        }
-                        .disabled(selectedSlot == nil)
-                        .padding(.top, 20)
-                    }
-                }
-                .padding()
-            }
-            .navigationTitle("Reschedule Appointment")
-            .navigationBarTitleDisplayMode(.inline)
-            .navigationBarItems(
-                leading: Button("Cancel") {
-                    presentationMode.wrappedValue.dismiss()
-                }
-            )
-            .onAppear {
-                selectedDate = initialDate
-                fetchDoctorDetails()
-                fetchSlots()
-            }
-            .alert(isPresented: $showConfirmation) {
-                Alert(
-                    title: Text(rescheduleSuccess ? "Success" : "Error"),
-                    message: Text(rescheduleSuccess ?
-                                  "Your appointment has been rescheduled successfully!" :
-                                  "Failed to reschedule appointment. Please try again."),
-                    dismissButton: .default(Text("OK")) {
-                        if rescheduleSuccess {
-                            onRescheduleComplete()
-                            presentationMode.wrappedValue.dismiss()
-                        }
-                    }
-                )
-            }
-        }
-    }
-    
-    private func formatDate(_ dateString: String) -> String {
-        let inputFormatter = DateFormatter()
-        inputFormatter.dateFormat = "yyyy-MM-dd"
-        
-        let outputFormatter = DateFormatter()
-        outputFormatter.dateStyle = .medium
-        
-        if let date = inputFormatter.date(from: dateString) {
-            return outputFormatter.string(from: date)
-        }
-        return dateString
-    }
-    
-    private func isSlotPassed(date: Date, timeString: String) -> Bool {
-        let now = Date()
-        let calendar = Calendar.current
-        
-        if !calendar.isDateInToday(date) {
-            return date < now
-        }
-        
-        let timeComponents = timeString.components(separatedBy: ":")
-        guard timeComponents.count >= 2,
-              let hour = Int(timeComponents[0]),
-              let minute = Int(timeComponents[1]) else {
-            print("Failed to parse time: \(timeString)")
-            return false
-        }
-        
-        let currentComponents = calendar.dateComponents([.hour, .minute], from: now)
-        let currentHour = currentComponents.hour ?? 0
-        let currentMinute = currentComponents.minute ?? 0
-        
-        if hour < currentHour {
-            return true
-        } else if hour == currentHour {
-            return minute <= currentMinute
-        } else {
-            return false
-        }
-    }
-    
-    private func fetchDoctorDetails() {
-        isLoading = true
-        errorMessage = nil
-        
-        guard let url = URL(string: "\(Constants.baseURL)/hospital/general/doctors/\(doctorId)/") else {
-            errorMessage = "Invalid URL"
-            isLoading = false
-            return
-        }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("Bearer \(UserDefaults.accessToken)", forHTTPHeaderField: "Authorization")
-        
-        URLSession.shared.dataTask(with: request) { data, response, error in
-            DispatchQueue.main.async {
-                isLoading = false
-                
-                if let error = error {
-                    errorMessage = error.localizedDescription
-                    return
-                }
-                
-                guard let httpResponse = response as? HTTPURLResponse else {
-                    errorMessage = "Invalid response"
-                    return
-                }
-                
-                guard (200...299).contains(httpResponse.statusCode), let data = data else {
-                    errorMessage = "Server returned status code \(httpResponse.statusCode)"
-                    return
-                }
-                
-                do {
-                    let response = try JSONDecoder().decode(PatientSpecificDoctorResponse.self, from: data)
-                    self.doctor = response
-                } catch {
-                    errorMessage = "Failed to decode response: \(error.localizedDescription)"
-                }
-            }
-        }.resume()
-    }
-    
-    private func fetchSlots() {
-        isLoading = true
-        
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
-        let dateString = dateFormatter.string(from: selectedDate)
-        
-        guard let url = URL(string: "\(Constants.baseURL)/hospital/general/doctors/\(doctorId)/slots/?date=\(dateString)") else {
-            errorMessage = "Invalid URL"
-            isLoading = false
-            return
-        }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("Bearer \(UserDefaults.accessToken)", forHTTPHeaderField: "Authorization")
-        
-        URLSession.shared.dataTask(with: request) { data, response, error in
-            DispatchQueue.main.async {
-                isLoading = false
-                
-                if let error = error {
-                    errorMessage = error.localizedDescription
-                    return
-                }
-                
-                guard let httpResponse = response as? HTTPURLResponse else {
-                    errorMessage = "Invalid response"
-                    return
-                }
-                
-                guard (200...299).contains(httpResponse.statusCode), let data = data else {
-                    errorMessage = "Server returned status code \(httpResponse.statusCode)"
-                    return
-                }
-                
-                do {
-                    let response = try JSONDecoder().decode([PatientSlotListResponse].self, from: data)
-                    self.slots = response.filter { !$0.is_booked }
-                    self.selectedSlot = nil
-                } catch {
-                    errorMessage = "Failed to decode response: \(error.localizedDescription)"
-                }
-            }
-        }.resume()
-    }
-    
-    private func rescheduleAppointment() {
-        guard let selectedSlot = selectedSlot else { return }
-        
-        isLoading = true
-        
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
-        let dateString = dateFormatter.string(from: selectedDate)
-        
-        let requestBody: [String: Any] = [
-            "date": dateString,
-            "slot_id": selectedSlot.slot_id
-        ]
-        
-        guard let url = URL(string: "\(Constants.baseURL)/hospital/general/appointments/\(appointmentId)/reschedule/") else {
-            errorMessage = "Invalid URL"
-            isLoading = false
-            return
-        }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("Bearer \(UserDefaults.accessToken)", forHTTPHeaderField: "Authorization")
-        
-        do {
-            request.httpBody = try JSONSerialization.data(withJSONObject: requestBody)
-        } catch {
-            errorMessage = "Failed to encode request: \(error.localizedDescription)"
-            isLoading = false
-            return
-        }
-        
-        URLSession.shared.dataTask(with: request) { data, response, error in
-            DispatchQueue.main.async {
-                isLoading = false
-                
-                if let error = error {
-                    errorMessage = error.localizedDescription
-                    showConfirmation = true
-                    rescheduleSuccess = false
-                    return
-                }
-                
-                guard let httpResponse = response as? HTTPURLResponse else {
-                    errorMessage = "Invalid response"
-                    showConfirmation = true
-                    rescheduleSuccess = false
-                    return
-                }
-                
-                if (200...299).contains(httpResponse.statusCode) {
-                    rescheduleSuccess = true
-                    showConfirmation = true
-                } else {
-                    errorMessage = "Server error: \(httpResponse.statusCode)"
-                    rescheduleSuccess = false
-                    showConfirmation = true
-                }
-            }
-        }.resume()
-    }
-}
-
 #Preview {
     HomePatient()
 }
